Extract Duplex transport
